requirements.txt
fastapi
uvicorn
torch
transformers
Pillow
requests
beautifulsoup4
